fix: use SelectAssetOverlay for overlay and create wrapper component for swap modal
